+++ src/lightning/pytorch/core/module.py
@@ -1519,6 +1519,13 @@ class LightningModule(
             **class** to call it instead of the :class:`LightningModule` instance, or a
             ``TypeError`` will be raised.
 
+        Note:
+            To ensure all layers can be loaded from the checkpoint, this function will call
+            :meth:`~lightning.pytorch.core.hooks.ModelHooks.configure_model` directly after instantiating the
+            model if this hook is overridden in your LightningModule. However, note that ``load_from_checkpoint`` does
+            not support loading sharded checkpoints, and you may run out of memory if the model is too large. In this
+            case, consider loading through the Trainer via ``.fit(ckpt_path=...)``.
+
         Example::
 
             # load weights without mapping ...