+++ src/lightning/pytorch/core/saving.py
@@ -38,6 +38,7 @@ from lightning.pytorch.accelerators import CUDAAccelerator, MPSAccelerator, XLAA
 from lightning.pytorch.utilities.imports import _OMEGACONF_AVAILABLE
 from lightning.pytorch.utilities.migration import pl_legacy_patch
 from lightning.pytorch.utilities.migration.utils import _pl_migrate_checkpoint
+from lightning.pytorch.utilities.model_helpers import is_overridden
 from lightning.pytorch.utilities.parsing import parse_class_init_keys
 from lightning.pytorch.utilities.rank_zero import rank_zero_warn
 
@@ -157,6 +158,9 @@ def _load_state(
     obj = cls(**_cls_kwargs)
 
     if isinstance(obj, pl.LightningModule):
+        if is_overridden("configure_model", obj):
+            obj.configure_model()
+
         # give model a chance to load something
         obj.on_load_checkpoint(checkpoint)